backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic-settings
python-docx
openai
azure-ai-documentintelligence
azure-core
